+++ src/studio/src/designer/frontend/packages/schema-editor/src/components/SchemaInspector.test.tsx
@@ -5,12 +5,7 @@ import { mount } from 'enzyme';
 import { act } from 'react-dom/test-utils';
 import { Autocomplete } from '@material-ui/lab';
 import { MenuItem } from '@material-ui/core';
-import SchemaInspector, {
-  isValidName,
-  isNameInUse,
-  isPathOnPropertiesRoot,
-  isPathOnDefinitionsRoot,
-} from './SchemaInspector';
+import SchemaInspector from './SchemaInspector';
 import { dataMock } from '../mockData';
 import { buildUISchema, resetUniqueNumber } from '../utils/schemaUtils';
 import type { ISchemaState, IUiSchemaItem } from '../types';
@@ -474,130 +469,7 @@ test('dispatches correctly when adding fields', () => {
   });
 });
 
-test('should not be possible to have two properties with the same name', () => {
-  const uiSchemaItems: IUiSchemaItem[] = [
-    {
-      path: '#/definitions/name',
-      type: 'object',
-      properties: [
-        {
-          path: '#/definitions/name/properties/child1',
-          type: 'string',
-          displayName: 'child1',
-        },
-        {
-          path: '#/definitions/name/properties/child2',
-          type: 'string',
-          displayName: 'child2',
-        },
-      ],
-      displayName: 'name',
-    },
-    {
-      path: '#/definitions/name2',
-      type: 'object',
-      properties: [
-        {
-          path: '#/definitions/name/properties/child1',
-          type: 'string',
-          displayName: 'child21',
-        },
-        {
-          path: '#/definitions/name/properties/child2',
-          type: 'string',
-          displayName: 'child22',
-        },
-      ],
-      displayName: 'name2',
-    },
-  ];
 
-  expect(
-    isNameInUse({
-      uiSchemaItems: uiSchemaItems,
-      parentSchema: uiSchemaItems[0],
-      path: '#/definitions/name',
-      name: 'name',
-    }),
-  ).toBe(false);
-  expect(
-    isNameInUse({
-      uiSchemaItems: uiSchemaItems,
-      parentSchema: uiSchemaItems[0],
-      path: '#/definitions/name/properties/child1',
-      name: 'child2',
-    }),
-  ).toBe(true);
-  expect(
-    isNameInUse({
-      uiSchemaItems: uiSchemaItems,
-      parentSchema: uiSchemaItems[0],
-      path: '#/definitions/name/properties/child2',
-      name: 'child3',
-    }),
-  ).toBe(false);
-  expect(
-    isNameInUse({
-      uiSchemaItems: uiSchemaItems,
-      parentSchema: null,
-      path: '#/properties/name',
-      name: 'name2',
-    }),
-  ).toBe(true);
-  expect(
-    isNameInUse({
-      uiSchemaItems: uiSchemaItems,
-      parentSchema: null,
-      path: '#/properties/name4',
-      name: 'name4',
-    }),
-  ).toBe(false);
-  expect(
-    isNameInUse({
-      uiSchemaItems: uiSchemaItems,
-      parentSchema: null,
-      path: '#/definitions/name',
-      name: 'name2',
-    }),
-  ).toBe(true);
-  expect(
-    isNameInUse({
-      uiSchemaItems: uiSchemaItems,
-      parentSchema: null,
-      path: '#/definitions/name4',
-      name: 'name4',
-    }),
-  ).toBe(false);
-});
-
-test('validates that the name only contains legal characters', () => {
-  expect(isValidName('Melding')).toBe(true);
-  expect(isValidName('melding')).toBe(true);
-  expect(isValidName('melding1')).toBe(true);
-  expect(isValidName('melding.1')).toBe(true);
-  expect(isValidName('melding_xyz')).toBe(true);
-  expect(isValidName('melding-xyz')).toBe(true);
-  expect(isValidName('1melding')).toBe(false);
-  expect(isValidName('_melding')).toBe(false);
-  expect(isValidName('.melding')).toBe(false);
-  expect(isValidName('melding%')).toBe(false);
-});
-
-test('should match if a properties based path is on root', () => {
-  expect(isPathOnPropertiesRoot('#/properties/prop1')).toBe(true);
-  expect(isPathOnPropertiesRoot('#/properties/prop1/properties/prop2')).toBe(
-    false,
-  );
-});
-
-test('should match if a definitions based path is on root', () => {
-  expect(isPathOnDefinitionsRoot('#/definitions/prop1')).toBe(true);
-  expect(isPathOnDefinitionsRoot('#/definitions/prop1/properties/prop2')).toBe(
-    false,
-  );
-  expect(isPathOnDefinitionsRoot('#/$defs/prop1')).toBe(true);
-  expect(isPathOnDefinitionsRoot('#/$defs/prop1/properties/prop2')).toBe(false);
-});
 
 test('dispatches correctly when changing type of combination', () => {
   const wrapper = mountWithId('#/definitions/allOfTest');
